refactor: 모듈화 작업
- 커스텀구분자와 일반구분자 코드를 따로 모듈화시킴
