+++ src/App.js
@@ -1,5 +1,7 @@
 import { Console } from '@woowacourse/mission-utils';
 import validateAnswer from './validator.js';
+import customDelimiter from './customDelimiter.js';
+import delimiter from './delimiter.js';
 class App {
 	async run() {
 		const answer = await Console.readLineAsync(
@@ -12,19 +14,12 @@ class App {
 			if (answer.trim() === '') {
 				Console.print(`결과: 0`);
 			} else if (answer.startsWith('//')) {
-				//커스텀 구분자
-				const delimiter = answer.substring(2, 3);
-				const numbers = answer.split('\\n')[1].split(delimiter);
-				const sum = numbers.map(Number).reduce((a, b) => a + b, 0);
-				Console.print(`결과 : ${sum}`);
+				customDelimiter(answer);
 			} else {
 				if (!validateAnswer(answer)) {
 					throw new Error(`[ERROR] : 사용자가 잘못된 값 입력`);
 				}
-				//일반 구분자
-				const numbers = answer.split(/[:,]/);
-				const sum = numbers.map(Number).reduce((a, b) => a + b, 0);
-				Console.print(`결과 : ${sum}`);
+				delimiter(answer);
 			}
 		} catch (error) {
 			throw new Error(error.message);
